requirements.txt
blinker==1.9.0
click==8.1.8
config==0.5.1
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
paho-mqtt==2.1.0
Werkzeug==3.1.3
